Fixed controls init extensions
